+++ user-service/src/main/java/com/arpangroup/user_service/processor/bonus/strategy/ReferralBonusStrategy.java
@@ -3,7 +3,9 @@ package com.arpangroup.user_service.processor.bonus.strategy;
 import com.arpangroup.user_service.entity.User;
 import com.arpangroup.user_service.entity.config.CalculationType;
 
+import java.math.BigDecimal;
+
 public interface ReferralBonusStrategy {
-    double calculateBonus(User user);
+    BigDecimal calculateBonus(User user);
     CalculationType getSupportedType();
 }
